requirements.txt
streamlit
python-dotenv
langchain
langchain_community
google-generativeai
langchain-google-genai 
PyPdf2
-e .